Updates to README.md
I'm proposing a few changes to the README.md file to make the setup cleaner. There were a few steps to edit a file, but the changes specified are already in the files that are checked in to master. If it is safe to assume that users will be building off the latest code, then there is no need to instruct them to change those items. I also put commands inside the `code` formatting syntax.